Function refactoring
